further label processing + parametrize
 Modifications qui seront validées :
	modifié :         ecospold2matrix.py
	modifié :         initialize_database.sql
	modifié :         performing_characterisation_matching.py

+++ performing_characterisation_matching.py
@@ -1,7 +1,4 @@
-
-# coding: utf-8
-
-# In[ ]:
+fullrun = False
 
 import ecospold2matrix as e2m
 import imp
@@ -18,7 +15,10 @@ import matlab_tools
 imp.reload(matlab_tools)
 
 project_name = 'testing'
-os.system('rm ' + project_name + '_characterisation.db')
+
+if fullrun:
+    os.system('rm ' + project_name + '_characterisation.db')
+
 
 parser=e2m.Ecospold2Matrix('/mnt/collection/current_Version_3.1_cutoff_ecoSpold02/', 
                            project_name,
@@ -26,19 +26,28 @@ parser=e2m.Ecospold2Matrix('/mnt/collection/current_Version_3.1_cutoff_ecoSpold0
                            verbose=False)
 self = parser
 c = self.conn.cursor()
+
+print("Getting ecoinvent labels in object")
 parser.get_labels()
-oldeco_path = '/home/bill/documents/arda/dev_arda_client/data/ecoinvent/2.2/Ecoinvent22_ReCiPe108_H.mat'
-matdict = scipy.io.loadmat(oldeco_path)
-a = np.array(matlab_tools.mine_nested_array(matdict['STR'], ''), dtype=object)
-self.STR_old = pd.DataFrame(a, columns=matlab_tools.mine_nested_array(matdict['STR_header'], '').squeeze().tolist())
 
+if fullrun:
+    print("initialize database")
+    parser.initialize_database()
+    print("Processing ecoinvent flows")
+    parser.process_ecoinvent_elementary_flows()
+    print("reading characterised flows")
+    parser.read_characterisation()
+    parser.populate_complementary_tables()
+    os.system('cp ' + project_name + '_characterisation.db start_characterisation.db')
+    print('DONE!!!')
+    try:
+        IPython.embed()
+    except:
+        pass
 
+parser.integrate_flows()
 
-
-parser.initialize_database()
-parser.process_ecoinvent_elementary_flows()
-parser.read_characterisation()
-parser.populate_complementary_tables()
-os.system('cp ' + project_name + '_characterisation.db start_characterisation.db')
-print('DONE!!!')
-IPython.embed()
+#oldeco_path = '/home/bill/documents/arda/dev_arda_client/data/ecoinvent/2.2/Ecoinvent22_ReCiPe108_H.mat'
+#matdict = scipy.io.loadmat(oldeco_path)
+#a = np.array(matlab_tools.mine_nested_array(matdict['STR'], ''), dtype=object)
+#self.STR_old = pd.DataFrame(a, columns=matlab_tools.mine_nested_array(matdict['STR_header'], '').squeeze().tolist())
